+++ src/App.tsx
@@ -1,14 +1,14 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 
 import HomePage from "./Pages/HomePage/index";
-import LoginPage2 from "./Pages/LoginPage/index";
+import LoginPage from "./Pages/LoginPage";
 
 const App = () => {
   return (
     <Router>
       <Routes>
         <Route path="/" element={<HomePage />} />
-        <Route path="/login" element={<LoginPage2 />} />
+        <Route path="/login" element={<LoginPage />} />
       </Routes>
     </Router>
   );
