+++ src/utils/const.ts
@@ -1,0 +1,3 @@
+const BASE_API: string = "http://localhost:3000";
+
+export { BASE_API };
